231016_1012

+++ ft_strmapi.c
@@ -15,9 +15,9 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char		*str;
-	size_t		s_len;
-	size_t		i;
+	char			*str;
+	size_t			s_len;
+	unsigned int	i;
 
 	i = 0;
 	s_len = ft_strlen(str);
